+++ activerecord/test/models/admin/user.rb
@@ -37,6 +37,7 @@ class Admin::User < ActiveRecord::Base
   store :json_data, accessors: [ :height, :weight ], coder: Coder.new
 
   store :json_data_empty, accessors: [ :is_a_good_guy ], coder: Coder.new
+  store :json_native_data, accessors: [ :city, :zip_code ], coder: JSON
 
   store_accessor :json_options, :enable_friend_requests